feat: send comments with post

+++ internal/store/storage.go
@@ -13,7 +13,7 @@ var (
 type Storage struct {
 	Posts interface {
 		Create(ctx context.Context, post *Post) error
-		GetByID(ctx context.Context, id int) (Post, error)
+		GetByID(ctx context.Context, id int64) (Post, error)
 	}
 	Users interface {
 		Create(ctx context.Context, user *User) error
